admin/requirements.txt
# =============================================================================
# Python зависимости для админ-панели
# =============================================================================

# FastAPI и ASGI сервер
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# База данных (асинхронный PostgreSQL)
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0

# Валидация и настройки
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Аутентификация
python-jose[cryptography]>=3.3.0
passlib>=1.7.4

# HTTP клиент
httpx>=0.26.0

# Утилиты
python-dateutil>=2.8.2
